+++ src/ImageSharp/Processing/Processors/PixelShading/PositionAwarePixelShaderProcessor.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0.
 
 using SixLabors.ImageSharp.PixelFormats;
-using SixLabors.ImageSharp.Processing.Delegates;
 using SixLabors.Primitives;
 
 namespace SixLabors.ImageSharp.Processing.Processors.PixelShading
